revise waiting/polling logic, reset probe frequency
* Go API/SDK: revise waiting/polling logic:
* start "small" and grow the probing interval incrementally
* allow callers to reset the latter (re: consecutive-idles logic)
* amend testing scripts and misc. CLI

+++ ais/test/scripts/cp-rmnode-rebalance.sh
@@ -13,7 +13,7 @@ trap cleanup EXIT INT TERM
 
 while true
 do
-  ## 1. start copying all
+  ## 1. start copying with an empty template (any prefix, any range)
   ais cp ais://src ais://dst --template ""
   sleep $((2 + RANDOM % 2))
 
@@ -22,10 +22,8 @@ do
   ais cluster add-remove-nodes start-maintenance $node -y ## triggers global rebalance
   sleep $((2 + RANDOM % 2))
 
-  ## 3. wait for both copying and rebalance to finish
+  ## 3. wait for the copy
   ais wait copy-objects
-  ais wait rebalance
-  ais wait copy-objects >/dev/null ## a no-op and won't take time
 
   ## 4. check the numbers
   res=$(ais ls ais://dst --no-headers | wc -l)
@@ -33,10 +31,18 @@ do
   res=$(ais ls ais://src --no-headers | wc -l)
   [[ $num == $res ]] || { echo "FAIL: source $num != $res"; exit 1; }
 
-  ## 5. activate and join back
+  ## 5. activate the (previously lost) node and join it back
+  ## (rebalance may be still running at this point - will be aborted and restarted)
   ais cluster add-remove-nodes stop-maintenance $node
 
   ## 6. cleanup and repeat
   ais wait rebalance
+
+  ## and check again
+  res=$(ais ls ais://dst --no-headers | wc -l)
+  [[ $num == $res ]] || { echo "FAIL: destination $num != $res"; exit 1; }
+  res=$(ais ls ais://src --no-headers | wc -l)
+  [[ $num == $res ]] || { echo "FAIL: source $num != $res"; exit 1; }
+
   ais bucket rm ais://dst -y
 done
